refactor: extract shared groove actions into useGrooveActions hook
- Create new useGrooveActions hook with shared logic for:
  - Note toggling (handleNoteToggle, handleSetNotes)
  - Measure manipulation (handleMeasureDuplicate, handleMeasureAdd,
    handleMeasureRemove, handleMeasureClear, handleClearAll)
  - Metadata updates (handleTitleChange, handleAuthorChange, handleCommentsChange)
- Refactor ProductionPage.tsx to use the new hook (~90 lines removed)
- Refactor GrooveEditor.tsx to use the new hook (~100 lines removed)
- Export NoteChange interface from the hook for batch note operations

Fixes #40
